Fix home page breakdown.
