+++ src/components/cart/CartProducts.js
@@ -1,40 +1,25 @@
-import { useEffect, useState } from "react";
+import {useEffect, useState} from "react";
 
-import { Button, Container, Grid, List, Typography } from "@mui/material";
+import {Button, Container, Grid, List, Typography} from "@mui/material";
 
 import CartProduct from "./CartProduct";
-import { getCartGames, getOrCreateCart } from "../../services";
-import { useAuth } from "../../context/AuthContext";
+import {getCartGames, getOrCreateCart} from "../../services";
+import {useAuth} from "../../context/AuthContext";
 
-export default function CartProducts() {
-    const [games, setGames] = useState([]);
-    const { getToken } = useAuth();
+export default function CartProducts({games, total}) {
+    const {getToken} = useAuth();
     const [showedGamesLimit, setShowedGamesLimit] = useState(12);
-    const [total, setTotal] = useState(0);
-
-    useEffect(() => {
-        async function fetchGames() {
-            const token = await getToken();
-            const gs = await getCartGames(token);
-            console.log("gs:", gs);
-            setGames(gs);
-            const gameIdToPrice = new Map(gs.map(g => [g.id, g.price]));
-            const cart = await getOrCreateCart(token);
-            const t = cart.games.map(g => g.amount * gameIdToPrice.get(g.id)).reduce((prev, curr) => prev + curr, 0);
-            setTotal(t);
-        }
-        fetchGames();
-    }, [])
 
     const renderProducts = games.map((product) => {
         return (
             <Grid item key={product.id} xs={2} sm={4} md={4} display="flex" flexDirection={'row'}
-                alignItems="center">
-                {<CartProduct product={product} />}
+                  alignItems="center">
+                {<CartProduct product={product}/>}
             </Grid>
         );
     }).slice(0, showedGamesLimit);
-    const onCheckout = () => { };
+    const onCheckout = () => {
+    };
 
     return (
         <Container id="products" maxWidth={"lg"}>
